add comments for procedure
