+++ confignode/src/test/java/org/apache/iotdb/confignode/procedure/impl/pipe/DropPipePluginProcedureTest.java
@@ -19,6 +19,7 @@
 
 package org.apache.iotdb.confignode.procedure.impl.pipe;
 
+import org.apache.iotdb.confignode.procedure.impl.pipe.plugin.DropPipePluginProcedure;
 import org.apache.iotdb.confignode.procedure.store.ProcedureFactory;
 import org.apache.iotdb.tsfile.utils.PublicBAOS;
